+++ data/cube/Cube.ts
@@ -227,6 +227,18 @@ export class Cube extends HoistBase {
         this._connectedViews.delete(view);
     }
 
+    /** Connect a View to this Cube for live updates. */
+    connectView(view: View) {
+        if (this.viewIsConnected(view)) return;
+
+        this._connectedViews.add(view);
+
+        // If the view is not up-to-date with the current cube data, then reload the view
+        if (view.info !== this.info) {
+            view.noteCubeLoaded();
+        }
+    }
+
     //-------------------
     // Data Loading API
     //-------------------